Modifs fichiers

+++ controllers/book.js
@@ -1,13 +1,18 @@
 const Book = require("../models/book");
+const fs = require("fs");
 
 exports.createBook = (req, res, next) => {
-    delete req.body._id;
-    const book = new Book({
-      ...req.body   //--> Raccourcis de title: req.body.title ... ça copie les champs du corps de la rquête et va le détaillé
-    });
-    book.save()
-      .then(() => res.status(201).json({ message: "Livre enregistré !" }))
-      .catch(error => res.status(400).json({ error }));  //--> c'est l'équivalent { error: error }
+  const bookObject = JSON.parse(req.body.book);
+  delete bookObject._id;
+  delete bookObject._userId;
+  const book = new Book({
+    ...bookObject,
+    userId: req.auth.userId,
+    imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
+  });
+  book.save()
+  .then(() => res.status(201).json({ message: "Livre enregistré !" }))
+  .catch(error => res.status(400).json({ error }));
 };
 
 exports.getOneBook = (req, res, next) => {
@@ -17,19 +22,43 @@ exports.getOneBook = (req, res, next) => {
 };
 
 exports.getAllBooks = (req, res, next) => {
-    Book.findOne({ _id: req.params.id })
+    Book.find() 
     .then(books => res.status(200).json(books))
     .catch(error => res.status(404).json({ error }));
 };
 
 exports.modifyBook = (req, res, next) => {
-  Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-  .then(() => res.status(200).json({ message: "Livre modifié !"}))
+  const bookObject = req.file ? {
+    ...JSON.parse(req.body.book),
+    imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
+  } : {...req.body};
+  delete bookObject._userId;
+  Book.findOne({_id: req.params.id})
+  .then((book) => {
+    if (book.userId != req.auth.userId) {
+      res.status(401).json({ message: "Non-autorisé" });
+    } else {
+      Book.updateOne({_id: req.params.id}, {...bookObject, _id: req.params.id})
+      .then(() => res.status(200).json({ message: "Livre modifié !" }))
+      .catch(error => res.status(401).json({ error }));
+    }
+  })
   .catch(error => res.status(404).json({ error }));
 };
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-  .then(() => res.status(200).json({ message: "Livre supprimé !"}))
-  .catch(error => res.status(404).json({ error }));
+  Book.findOne({_id: req.params.id})
+  .then(book => {
+    if (book.userId != req.auth.userId) {
+      res.status(401).json({ message: "Non-autorisé" });
+    } else {
+      const filename = book.image.Url.split("/images/")[1];
+      fs.unlink(`images/${filename}`), () => {
+        Book.deleteOne({_id: req.params.id})
+        .then(() => {res.status(200).json({ message: "Livre suppromé !" })})
+        .catch(error => res.status(401).json({ error}));
+      }
+    }
+  })
+  .catch(error => res.status(500).json({ error }));
 };
